Saparate sample app screens from App.js
